requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.6.3
Brotli
cachetools==5.3.2
certifi
cffi
charset-normalizer
click
colorama
cryptography
Flask==3.0.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna
importlib-metadata
itsdangerous==2.1.2
Jinja2
joblib==1.3.2
keras==2.13.1
Keras-Preprocessing
libclang==16.0.6
mkl-fft
mkl-random
mkl-service==2.4.0
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging
platformdirs
pooch
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser
pyOpenSSL
PySocks
python-dateutil==2.8.2
regex==2023.10.3
requests
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow~=2.13.0
tqdm==4.66.1
typing_extensions==4.5.0
urllib3
Werkzeug==3.0.1
win-inet-pton
wrapt==1.15.0
zipp
python-dotenv~=1.0.1
gunicorn~=22.0.0